# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Soroban Pulse API",
    "description": "Indexes Soroban smart contract events on the Stellar network.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/events/contract/{contract_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "get_events_by_contract",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "description": "Stellar contract ID (56-char, starts with C)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Results per page, 1–100 (default: 20)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor for keyset pagination (returned as next_cursor)",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order: asc (oldest first) or desc (newest first, default)",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "enum": ["asc", "desc"]
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort column: ledger (default), timestamp, or created_at",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "enum": ["ledger", "timestamp", "created_at"]
            }
          },
          {
            "name": "from_ledger",
            "in": "query",
            "description": "Return events at or after this ledger",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "to_ledger",
            "in": "query",
            "description": "Return events at or before this ledger",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events for the given contract"
          },
          "400": {
            "description": "Invalid contract_id format or ledger range"
          }
        }
      }
    },
...
  },
...
}